Add support for the TRANSP and CLASS properties
Per the CalDAV spec, setting TRANSP to TRANSPARENT makes a time appear free
in the free/busy schedule.
Default is OPAQUE

If CLASS is PUBLIC, show the summary.
If CONFIDENTIAL, don't even include it.
If PRIVATE, hide the summary.
Default is PUBLIC!
this is a shitty default due to the behaviour of some clients.